[DESK-73] Place the Sync to perform on HOME view CLI
Placed the Sync to perform on HOME view CLI
